api/requirements.txt
# API dependencies
Flask==3.1.2
Werkzeug==3.1.3
requests==2.32.5
python-dotenv==1.1.1
firebase-admin==7.1.0
grpcio>=1.33.2,<2.0.0
Pillow==11.3.0
colorgram.py==1.2.0
markupsafe==3.0.3
gunicorn==23.0.0
profanityfilter==2.1.0

# Test dependencies for CI/CD
pytest==8.4.2
pytest-mock==3.15.1
pytest-cov==6.0.0
